requirements.txt
qwen-vl-utils
pillow
lmdeploy
torch
tqdm
transformers
ultralytics
